+++ lib/src/video_quality.dart
@@ -1,3 +1,8 @@
 part of flutter_video_compress;
 
-enum VideoQuality { LowQuality, MediumQuality, HighestQuality }
+enum VideoQuality { 
+  DefaultQuality, 
+  LowQuality, 
+  MediumQuality, 
+  HighestQuality
+}